requirements.txt
python-decouple
swibots
requests